better documentation, restructure
Shuffle stuff around and rename identifiers for consistency, or else
I'll go mad in just a day forgetting this. Need to digitize the sketches
on paper sooner or later, too. Making the outer bit translucent is very
useful for debugging.

The lock clearance doesn't seem consistent... A real constraint-based
CAD with an understanding of surfaces would be so much better for this.