initial commit for device-side input adapter
The machine itself does not support an input hose for the hot side but
it's easily hacked with a cardboard box and some adapters. This will be
the adapter that goes to the box on the machine's hot-side input hole.

I've got an annoying hose there already. The better hose I've got is too
different to fit the existing bits and the shop doesn't stock a matching
adapter.

Hey openscad, why no combination of a linear extrude and a rotate
extrude? Apparently people have made screw libs already but let's make a
custom one to have some fun. The linear extrude twists in the wrong axis
for stuff like this and rotate extrude doesn't support z translation.