some screw planning
Openscad doesn't include a good way to do screw-like things, so let's
roll our own with a single polyhedron. Coming up with a diy
rotate_extrude() with extra support to translate along the way would be
a natural way to do this, or linear_extrude() with the cross section in
a different angle. Overdoing a general operation to use an extrusion
track and a cross section must be perfect. Unfortunately functions
aren't first-class things in the language so a vertex transform function
won't do.