fix screw winding
The back face needs a mirrored orientation to be consistent.

Fix docs: the given face / cross section has to be counterclockwise
because the screw is constructed that way: when looking at the section
flat from the z axis, the "back side" of it is seen as it grows that way
with the right hand rule.