Final docs on the input flange adapter
The flanged nut bit for the input hose is now printed. It's not perfect
but it's fine.

The threads were a bit tricky to clean up from support material (which
probably was necessary due to the significant overhang). An even
side/body thickness for the threads would have mostly prevented this,
but it would have caused perpendicular steps on the printed layers that
might interfere with the ease of installation. Probably not though if
the steel wire inside the hose would be just twisted a bit from the end.