make the thread tip less blunt for easier install
This does not matter anymore for the existing object I have, but the
other end of the hose will receive the same treatment. Figure out a nice
triangle from the end tip on the outer thread surface to the side of the
inner thread surface to meet it tangentially. Use the triangle to cut a
bit out of the threads.

Constructing the threads with this "ease-in" built in would be more
robust and more general but that would need more complex modifications
to the thread generation, for example by cutting the cross section more
and more towards the tip.