initial input hose collar object
This goes to the other end of the input hose, facing the window panel.
Now there's just this, not one for the hose and one for the panel.

TODO the outdoor part and locking mechanism. Maybe a manual knob to lock
the flaps for rain because they need to angle inwards along the airflow.